refactor: Auth Application 서비스 레이어 업데이트 #323
- AuthService: 새로운 도메인 모델 적용
- UserProviderAppender/Reader: 도메인 객체 기반으로 리팩토링
- 기존 UserProviderRepository 삭제 (Infrastructure로 이동)
- 순수 도메인 로직과 인프라 로직 분리 완료
